Rendering OUT instead of level for winners in high score table
Some code unification
