+++ client/views/mini-scores-view.js
@@ -5,10 +5,9 @@ define([
     "underscore",
     "backbone",
     "dispatcher",
-    "dataIO/send-generic",
     "views/view-activation-helpers",
-    'moment'
-], function ($, _, Backbone, dispatcher, send, activate, Moment) {
+    "views/score-table-cells"
+], function ($, _, Backbone, dispatcher, activate, TableCells) {
 
     var HighScoresView = Backbone.View.extend({
 
@@ -23,43 +22,6 @@ define([
         initialize: function() {
             this.listenTo(this.model, "add", this.render);
             this.listenTo(this.model, "change", this.render);
-
-            var self = this;
-
-            var WatchGameUriCell = Backgrid.UriCell.extend({
-
-                events : {
-                    "click #watch-game" : "watchGame"
-                },
-
-                watchGame: function(event){
-                    event.preventDefault();
-
-                    var gameId = $(event.target).data("gameid");
-                    var gameDescription = $(event.target).data("gamedescription");
-
-                    send("brogue", "recording", {recording: gameId});
-                    dispatcher.trigger("recordingGame", {recording: gameDescription});
-                    self.goToConsole();
-                },
-
-                render: function () {
-                    this.$el.empty();
-                    var rawValue = this.model.get(this.column.get("name"));
-                    var formattedValue = this.formatter.fromRaw(rawValue, this.model);
-                    if(formattedValue) {
-                        this.$el.append($("<a>", {
-                            href: '#brogue',
-                            title: this.model.title,
-                            id: 'watch-game',
-                            "data-gameid": formattedValue,
-                            "data-gamedescription": this.model.get("username") + "-" + this.model.get("seed") + "-" + self.formatDate(this.model.get("date"))
-                        }).text("Watch game"));
-                    }
-                    this.delegateEvents();
-                    return this;
-                }
-            });
 
             this.grid = new Backgrid.Grid({
                 columns: [
@@ -84,7 +46,7 @@ define([
                     }, {
                         name: "level",
                         label: "Level",
-                        cell: "integer",
+                        cell: TableCells.levelCell,
                         sortable: false,
                         editable: false
                     }, {
@@ -102,7 +64,7 @@ define([
                     }, {
                         name: "recording",
                         label: "Recording",
-                        cell: WatchGameUriCell,
+                        cell: TableCells.watchGameUriCell,
                         sortable: false,
                         editable: false
                     }],
@@ -164,11 +126,7 @@ define([
         goToConsole : function(){
             activate.console();
             dispatcher.trigger("showConsole");
-        },
-
-        formatDate: function(date) {
-            return Moment(date).format('MMMM Do YYYY, h:mm:ss a');
-        },
+        }
     });
 
     return HighScoresView;
